add basic import of predicted values

+++ cz.mzk.recordmanager.server/src/main/resources/sql/recordmanager-create-tables-derby.sql
@@ -317,6 +317,6 @@ CREATE TABLE classifier (
   id                   DECIMAL(10) PRIMARY KEY,
   harvested_record_id  DECIMAL(10),
   value                VARCHAR(255),
-  relevance            DECIMAL(30),
+  relevance            REAL,
   CONSTRAINT classifier_fk   FOREIGN KEY (harvested_record_id) REFERENCES harvested_record(id)
 );
