+++ cz.mzk.recordmanager.server/src/main/resources/sql/recordmanager-liquibase.sql
@@ -1277,7 +1277,7 @@ CREATE TABLE classifier (
   id                   DECIMAL(10) PRIMARY KEY,
   harvested_record_id  DECIMAL(10),
   value                VARCHAR(255),
-  relevance            DECIMAL(30),
+  relevance            REAL,
   FOREIGN KEY (harvested_record_id) REFERENCES harvested_record(id) ON DELETE CASCADE
 );
 CREATE INDEX classifier_harvested_record_idx ON classifier(harvested_record_id);
